@@ -1,4 +1,4 @@
-package com.sno.testit;
+package ${groupId};
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
